refactor: Rename component folder timeline to feed
